added panel extraction logic
